moved FileIcon to new folder; fixed lint issues; added FileIcon to export
