+++ src/components/Icon/Custom/ExclamationCircle.tsx
@@ -1,4 +1,4 @@
-import { type Icon } from './types';
+import { type Icon } from '../types';
 
 export const ExclamationCircle: Icon = props => {
   const size = props.size ?? 24;
