+++ src/components/Icon/Custom/SquareRoundedChevronDownFilled.tsx
@@ -1,4 +1,4 @@
-import { type Icon } from './types';
+import { type Icon } from '../types';
 
 export const SquareRoundedChevronDownFilled: Icon = props => {
   return (
