+++ src/components/Icon/Custom/index.ts
@@ -1,13 +1,3 @@
 export * from './ExclamationCircle';
 export * from './SquareRoundedChevronDownFilled';
 export * from './GameIconsPerson';
-export * from './FileIcon';
-export * from './FileIcon/CSV';
-export * from './FileIcon/DOC';
-export * from './FileIcon/DOCX';
-export * from './FileIcon/PDF';
-export * from './FileIcon/PPT';
-export * from './FileIcon/PPTX';
-export * from './FileIcon/TXT';
-export * from './FileIcon/XLS';
-export * from './FileIcon/XLSX';
