--- src/components/Icon/Custom/types.ts
@@ -1,8 +1,0 @@
-import { type FC, type SVGAttributes } from 'react';
-
-export interface IconProps extends SVGAttributes<SVGElement> {
-  color?: string;
-  size?: string | number;
-}
-
-export type Icon = FC<IconProps>;
