+++ src/components/Icon/FileIcon/DOC.tsx
