+++ src/components/Icon/FileIcon/PPTX.tsx
